requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==37.0.4
Django==4.0.6
django-cors-headers==3.13.0
django-rest-knox==4.2.0
djangorestframework==3.13.1
idna==3.3
pycparser==2.21
pytz==2022.1
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.12
